requirements.txt
langgraph
langchain-groq
langchain-core
langgraph-checkpoint-postgres
psycopg[binary]
deepeval
python-dotenv
fastapi
uvicorn
pydantic